feature: remove parm customizations
* parm now functions as an infinite loading app
* current is now displayed

I will need to make a larger refactor if I want to change
app behavior. The hack worked fine for my initial usecase,
but the application is hard to reason about when
it's designed as a choose-your-own adventure but
used as a blog.

Realistically, I should split components into smaller units
and combine them in ways that create each behavior, instead of
overloading the adventure mode for each.